wip: refactor instance mapping into new component
To display one instance type per step on full screen.
